fix(auth): fix custom claims linking for extra fields
The initial draft assumed that the `getUserByCustomClaims($claims)` function would receive the full userinfo data as received from the oauth provider as its `$claims` parameter.
But since the linking happens in the overlaying oauth plugin instead of in the service-specific oauth plugins, the userinfo is never available in its current context.
This commit works around this issue by also adding special 'extras' nested array handling for the claim field names.

Ultimately, the custom claims matching structure should be refactored and moved into the service-specific plugins, possibly with the use of an abstract/default `getUserByCustomClaims` method that each service can override as it sees fit.